+++ src/games/cowpoke/components/Feed.tsx
