+++ src/games/cowpoke/components/LootContainer.tsx
