+++ src/games/cowpoke/components/UseGameData.tsx
