+++ src/games/game-of-life/components/UseGameData.tsx
