+++ src/games/game-of-life/components/UseSettings.tsx
