+++ src/games/game-template/components/GameComponent.tsx
@@ -1,36 +1,25 @@
 "use client";
 
 import { useEffect, useState, useRef } from "react";
-import "@/src/games/better-boids/styles/game.css";
-import GameInfoContainer from "@/src/components/GameInfoContainer";
-import SettingsContainer from "@/src/games/better-boids/SettingsContainer";
+import "@/src/games/game-template/styles/game.css"; // FIXME: UPDATE THIS PATH TO THE GAME-SPECIFIC CSS
 import {
   loadPhaserScriptThenGame,
   cleanupPhaserGame,
 } from "@/src/utils/phaser-loading";
-import { ContentDataProps } from "@/src/types/data-props";
 import GameLoadingScreen from "@/src/components/GameLoadingScreen";
+import { ContentDataProps } from "@/src/types/data-props";
+import GameInfoContainer from "@/src/components/GameInfoContainer";
+import SettingsContainer from "@/src/games/game-template/components/SettingsContainer"; // FIXME: UPDATE THIS PATH TO THE GAME-SPECIFIC SETTINGS CONTAINER
 
+// FIXME: UPDATE THIS TO THE GAME-SPECIFIC GAME INFO
 export const gameInfoData: ContentDataProps[] = [
   {
     type: "h1",
-    text: "Better Boids",
+    text: "<Game Name>",
   },
   {
     type: "paragraph",
-    text: 'The <a href="https://en.wikipedia.org/wiki/Boids" target="_blank" rel="noopener noreferrer">Boids algorithm</a>, devised by Craig Reynolds, mimics the flocking behavior seen in birds and other animals. In general, Boids follow three rules:',
-  },
-  {
-    type: "list",
-    items: [
-      "<b>Alignment:</b> Boids try to align their direction with other nearby Boids.",
-      "<b>Cohesion:</b> Boids move towards the average position of nearby Boids.",
-      "<b>Separation:</b> Boids avoid crowding near other Boids.",
-    ],
-  },
-  {
-    type: "paragraph",
-    text: "From these three simple rules, complex emergent behavior and intricate patterns can arise. This little game is an attempt to display the beauty in the Boids algorithm, while expanding on it with novel concepts where applicable.",
+    text: "Fill in information about the game here",
   },
 ];
 
@@ -47,7 +36,8 @@ const GameComponent: React.FC = () => {
   };
 
   useEffect(() => {
-    document.body.classList.add("better-boids-game-background");
+    // FIXME: UPDATE THIS TO THE GAME-SPECIFIC BACKGROUND CLASS
+    document.body.classList.add("game-name-game-background");
 
     const loadPhaserGame = async () => {
       if (!window.Phaser) {
@@ -67,8 +57,9 @@ const GameComponent: React.FC = () => {
       }
 
       try {
+        // FIXME: UPDATE IMPORT PATH FOR THE GAME-SPECIFIC SCENE
         const { MainGameScene } = await import(
-          "@/src/games/better-boids/scenes/main-game-scene"
+          "@/src/games/game-template/scenes/main-game-scene"
         );
 
         const gameConfig: Phaser.Types.Core.GameConfig = {
@@ -113,7 +104,7 @@ const GameComponent: React.FC = () => {
       {/* Loading Screen */}
       {isLoading && (
         <GameLoadingScreen
-          coverImage="/webps/games/better-boids-cover.webp"
+          coverImage="/webps/games/better-boids-cover.webp" // FIXME: UPDATE THIS TO THE GAME-SPECIFIC COVER IMAGE
           onFadeOutComplete={handleFadeOutComplete}
         />
       )}
