+++ src/games/game-template/components/SettingsContainer.tsx
