+++ src/types/data-props.ts
@@ -1,24 +1,11 @@
-export interface GameDataProps {
+export interface GameMetadataProps {
   title: string;
   description: string;
   logoImageUrl: string;
   imageAlt: string;
 }
 
-export interface ContentDataProps {
-  type: "h1" | "h2" | "h3" | "paragraph" | "list";
-  text?: string;
-  items?: string[];
-}
-
-export interface StoryDataContentProps {
-  content: string;
-  fontStyle: "normal" | "italic" | "bold";
-  textAlign: "left" | "center" | "right" | "justify";
-  splitParagraphs?: boolean;
-}
-
-export interface StoryDataProps {
+export interface StoryMetadataProps {
   title: string;
   subtitle: string;
   date: string;
@@ -26,5 +13,18 @@ export interface StoryDataProps {
   imageWidth: number;
   imageHeight: number;
   artContent?: boolean;
-  body: StoryDataContentProps[];
+  body: StoryContentDataProps[];
+}
+
+export interface StoryContentDataProps {
+  content: string;
+  fontStyle: "normal" | "italic" | "bold";
+  textAlign: "left" | "center" | "right" | "justify";
+  splitParagraphs?: boolean;
+}
+
+export interface ContentDataProps {
+  type: "h1" | "h2" | "h3" | "paragraph" | "list";
+  text?: string;
+  items?: string[];
 }
